+++ src/App.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from 'styled-components';
 import defaultTheme from 'design/theme';
 import GlobalStyle from 'design/GlobalStyle';
 import { BrowserRouter } from 'react-router-dom';
-import RoutePage from 'RoutePage';
+import RoutePage from 'routes/RoutePage';
 import Header from 'components/header/Header';
 import NavigationBar from 'components/navigationBar/NavigationBar';
 import Footer from 'components/footer/Footer';
